server manager factory params changed
